+++ Sample.h
@@ -13,22 +13,22 @@ class Sample {
 public:
     float latitude;
     float longitude;
-    int waterDepth;
+    float waterDepth;
     float obsvnTop;
     float obsvnBot;
-    int gravel;
-    int sand;
-    int mud;
-    int clay;
+    float gravel;
+    float sand;
+    float mud;
+    float clay;
     float grainSize;
     float sorting;
     // We should consider converting this into it's separate components
     string munslColr;
     float orgCarbn;
-    int porosity;
+    float porosity;
 
     Sample();
-    Sample(float latitude, float longitude, int waterDepth, float obsvnTop, float obsvnBot, int gravel, int sand, int mud, int clay, float grainSize, float sorting, string& munslColr, float orgCarbn, int porosity);
+    Sample(float latitude, float longitude, float waterDepth, float obsvnTop, float obsvnBot, float gravel, float sand, float mud, float clay, float grainSize, float sorting, string& munslColr, float orgCarbn, float porosity);
 };
